feat e refactor: organização dos arquivos da api Python e ajustes nas consultas
