+++ src/main/chatbot/app/config/database_config.py
@@ -1,6 +1,6 @@
 import psycopg2
 
-from app.config import db_host, db_port, db_database, db_user, db_password
+from app.config.config import db_host, db_port, db_database, db_user, db_password
 
 def get_conexao():
     return psycopg2.connect(
